finished porting get_controller
