+++ backend/src/controllers/mod.rs
@@ -1,2 +1,1 @@
-pub mod get_controller;
 pub mod upload_controller;
