+++ backend/src/main.rs
@@ -4,6 +4,7 @@ mod utils;
 mod delete_service;
 mod database;
 mod routes;
+mod types;
 
 use crate::routes::configure_routes;
 use database::aws_service::AWSService;
