+++ backend/src/types/content.rs
@@ -1,0 +1,6 @@
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+pub struct ContentQuery {
+    pub compress: Option<bool>,
+}
